// Projeto/backend/src/models/OperationType.ts
export enum OperationType {
	LOGIN = 'login',
	GET_USER = 'getUser',
	GET_USERS = 'getUsers',
	CREATE_USER = 'createUser',
	UPDATE_USER = 'updateUser',
	DELETE_USER = 'deleteUser',
	CREATE_CHALLENGE = 'createChallenge',
	AUTH_CODE_GRANT = 'authCodeGrant',
	REFRESH_TOKEN = 'refreshToken',
	AUTHORIZE_URL = 'authorizeUrl',
}